feat(avatar-dropdown): introduced avatar dropdown component
